clean configuration for build
